refactor: GlobalErrorInfoModal -> GlobalErrorInfoComponent
추후 모달이 아닌 다른 방식으로서의 확장성을 고려
