+++ fourtoon-cookie/src/components/global/GlobalError/GlobalErrorInfoComponent/GlobalErrorInfoComponent.styled.ts
